book_store_backend/pyproject.toml
[tool.poetry]
name = "book-store"
version = "0.1.0"
description = "Book store server"
authors = ["thinpyai <thinpyaiwin@gmail.com>"]
readme = "README.md"
packages = [
    { include = "*", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.103.0"
uvicorn = {extras = ["standard"], version = "0.17.6"}
SQLAlchemy-Utils = "0.38.2"
alembic = "1.7.7"
strawberry-graphql = {extras = ["fastapi"], version = "0.109.1"}
requests = "2.27.1"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.0"
autopep8 = "^1.7.0"
pylint = "^2.15.0"
pytest = "^7.1.0"
pytest-mock = "^3.10.0"
pytest-asyncio = "^0.19.0"
pytest-cov = "^4.0.0"
coverage = "^6.5.0"
pywin32 = "^304"
freezegun = "^1.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
